Update getAnnotation.R
I was getting errors that the FTP download was not completing. The same as this issue from 2 years ago: https://github.com/BgeeDB/BgeeDB_R/issues/27. I found this was because of how the Bgee database and R package is set up now, the path that is set to where files are downloaded and extracted (pathToData), is the same path that line 76 of getAnnotation.R script deletes by unlinking (dirname(myData)). This removes all downloaded BGee data when running and therefore returns an error. I guess this line was left over from a previous set up. I think just removing will be a fix with no problems.